Split mesh redistribution into a separate branch
Remove firedrake/redist.py and the RedistributedMeshTransfer-based
redistribution support that was threaded through MeshHierarchy,
AdaptiveMeshHierarchy, and refine_marked_elements. This functionality
moves to a follow-up branch/PR built on top of this one.

make_unoverlapped_dm, the one redist.py helper that MeshHierarchy uses
unconditionally (independent of redistribution), is relocated into
firedrake/mg/mesh.py.
